+++ logger.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"context"
 	"encoding/hex"
+	"fmt"
 	"io/ioutil"
 	"log"
 
@@ -28,11 +29,11 @@ type Logger struct {
 	debug bool
 	// client ids
 	maxLen uint
+	remote string
 }
 
-func (l *Logger) SetDebug(v bool) {
-	l.debug = v
-}
+func (l *Logger) SetDebug(v bool)    { l.debug = v }
+func (l *Logger) SetRemote(v string) { l.remote = v }
 
 // SetMaxIDLen configures the logger to trim the client id to number of
 // characters. Use 0 to not trim.
@@ -44,18 +45,25 @@ func (l *Logger) SetMaxIDLen(max uint) {
 // mq.ConnAck.AssignedClientID.
 func (f *Logger) In(next Handler) Handler {
 	return func(ctx context.Context, p mq.Packet) error {
-		if p, ok := p.(*mq.ConnAck); ok {
+		var msg string
+		switch p := p.(type) {
+		case *mq.ConnAck:
 			if v := p.AssignedClientID(); v != "" {
 				f.SetLogPrefix(v)
 			}
+		case *mq.Connect:
+			msg = fmt.Sprintf("in  %v %s", p, f.remote)
+		}
+		if msg == "" {
+			msg = fmt.Sprint("in  ", p)
 		}
 		// double spaces to align in/out. Usually this is not advised
 		// but in here it really does aid when scanning for patterns
 		// of packets.
 		if f.debug {
-			f.Print("in  ", p, "\n", dumpPacket(p))
+			f.Print(msg, "\n", dumpPacket(p))
 		} else {
-			f.Print("in  ", p)
+			f.Print(msg)
 		}
 		err := next(ctx, p)
 		if err != nil {