+++ changelog.md
@@ -5,6 +5,10 @@ documented in this file.
 The format is based on http://keepachangelog.com/en/1.0.0/
 and this project adheres to http://semver.org/spec/v2.0.0.html.
 
+## [unreleased]
+
+- Logger logs optional remote addr on mq.Connect
+
 ## [0.4.0] 2022-11-25
 
 - Use pink gopher in logo